+++ src/Services/RuleConverters/Behaviours/HasLength.php
@@ -3,16 +3,17 @@
 namespace Somnambulist\Bundles\ApiBundle\Services\RuleConverters\Behaviours;
 
 use Somnambulist\Bundles\ApiBundle\Services\RuleConverters;
+use function array_key_exists;
 use function in_array;
 
 trait HasLength
 {
-    protected function hasLength(array $schema, RuleConverters $converters): bool
+    protected function hasLength(array $schema, array $rules): bool
     {
         return
             in_array($schema['type'], ['array', 'string'])
             ||
-            $converters->has('digits') || $converters->has('digits_between')
+            array_key_exists('digits', $rules) || array_key_exists('digits_between', $rules)
         ;
     }
 }