+++ src/Services/RuleConverters/MinConverter.php
@@ -15,7 +15,7 @@ class MinConverter extends AbstractRuleConverter
 
     public function apply(array $schema, string $rule, string $params, array $rules): array
     {
-        $schema[$this->hasLength($schema, $this->converters) ? 'minLength' : 'minimum'] = ctype_digit($params) ? (int)$params : (float)$params;
+        $schema[$this->hasLength($schema, $rules) ? 'minLength' : 'minimum'] = ctype_digit($params) ? (int)$params : (float)$params;
 
         return $schema;
     }